logon登录修改为session
logon登录修改为session

+++ syglWeb/slip/Profile/Prj/DeleteMb.ashx.cs
@@ -6,6 +6,7 @@ namespace syglWeb.slip.Profile.Prj
 {
     /// <summary>
     /// DeleteMb 的摘要说明
+    /// 13-1-1 15:57 用户权限验证已添加
     /// </summary>
     public class DeleteMb : IHttpHandler
     {
@@ -15,6 +16,7 @@ namespace syglWeb.slip.Profile.Prj
 
             int status = 0;
             string msg = "";
+
             string _mbID = context.Request["mbID"];
             if (_mbID == "" || _mbID == null)
             {
@@ -23,22 +25,32 @@ namespace syglWeb.slip.Profile.Prj
             }
             else
             {
+                
                 SRDel.SRSql srSql = new SRDel.SRSql();
                 try
                 {
-
-                    srSql.conn.Open();
-                    srSql.cmd.CommandText = "delete from topicMbTb where topicMbID= @topicMbID";
-                    srSql.cmd.Parameters.AddWithValue("@topicMbID", Convert.ToInt32(_mbID));
-                    srSql.cmd.ExecuteNonQuery();
-                    srSql.conn.Close();
+                    if (context.Session["userID"] == null)
+                    {
+                        status = 1;
+                        msg = "您未登录或登录超时！";
+                    }
+                    else
+                    {
+                        srSql.conn.Open();
+                        srSql.cmd.CommandText = "delete from topicMbTb where topicMbID= @topicMbID";
+                        srSql.cmd.Parameters.AddWithValue("@topicMbID", Convert.ToInt32(_mbID));
+                        srSql.cmd.ExecuteNonQuery();
+                        srSql.conn.Close();
+                    }
                 }
-                catch(Exception ex){
+                catch (Exception ex)
+                {
                     srSql.conn.Close();
                     status = 1;
                     msg = ex.ToString();
                 }
-                finally{
+                finally
+                {
                     srSql.conn.Close();
 
                     context.Response.AddHeader("Content-Type", "text/html; charset=UTF-8");
@@ -46,7 +58,8 @@ namespace syglWeb.slip.Profile.Prj
                     context.Response.End();
                 }
             }
-             
+
+
         }
 
         public bool IsReusable
